rename devices.p_layouts to layout rename simulator to simulators rename parser to parsers
