+++ hades/parsers/map.py
