+++ tests/test_generate/test_generate.py
@@ -1,6 +1,6 @@
 import os
 from hades.main import generate_cli
-from hades.parser.map import get_number
+from hades.parsers.map import get_number
 from os import chdir
 import yaml
 import pytest
