+++ tests/test_parser/test_tlef.py
@@ -1,5 +1,5 @@
 import pytest
-from hades.parser.tlef import load_tlef, get_metal, get_via
+from hades.parsers.tlef import load_tlef, get_metal, get_via
 from hades.techno import load_pdk
 from os.path import join, dirname, isdir
